+++ src/app/(smal)/auth/verify/page.tsx
@@ -1,29 +1,43 @@
 import { Suspense } from "react";
-import { VerifyForm } from "./page.client";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle
-} from "@/components/ui/card";
+import { CardBase } from "@/components/cardBase";
+import { EmailParam } from "./page.server";
+import { ResendEmailButton } from "./page.client";
+
+export const experimental_ppr = true;
 
 const Page = () => {
   return (
-    <div className="flex flex-col gap-6">
-      <Card className="border-0 shadow-none">
-        <CardHeader>
-          <CardTitle>Verify Your Email</CardTitle>
-          <CardDescription>
-            Please enter the verification code sent to your email.
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <Suspense>
-            <VerifyForm />
-          </Suspense>
-        </CardContent>
-      </Card>
+    <div>
+      <CardBase
+        title="Check your email"
+        description="We've sent a verification link to your email address."
+        content={
+          <div className="flex flex-col gap-6">
+            <div className="grid gap-3">
+              <p>We've sent a verification link to:</p>
+              <p className="text-blue-600 font-medium">
+                <Suspense>
+                  <EmailParam />
+                </Suspense>
+              </p>
+            </div>
+            <div className="grid gap-3">
+              <p>
+                Click the link in the email to verify your account and complete
+                your registration.
+              </p>
+              <p className="text-sm text-muted-foreground">
+                Don't forget to check your spam folder if you don't see the
+                email.
+              </p>
+            </div>
+            <div className="grid gap-3">
+              <p>Didn't receive the email?</p>
+              <ResendEmailButton />
+            </div>
+          </div>
+        }
+      />
     </div>
   );
 };
